Move ContentLOader to application/content and GitHandler to application/orchestration
