+++ src/main/groovy/com/cloudogu/gitops/Feature.groovy
@@ -2,8 +2,8 @@ package com.cloudogu.gitops
 
 import static com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy.RepoType
 
+import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.features.git.GitHandler
 import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
 import com.cloudogu.gitops.utils.AirGappedUtils
 import com.cloudogu.gitops.utils.FileSystemUtils
