+++ src/main/groovy/com/cloudogu/gitops/application/content/ContentLoader.groovy
@@ -1,13 +1,14 @@
-package com.cloudogu.gitops.features
+package com.cloudogu.gitops.application.content
 
 import static com.cloudogu.gitops.config.Config.ContentRepoType
 import static com.cloudogu.gitops.config.Config.ContentSchema.ContentRepositorySchema
 
 import com.cloudogu.gitops.Feature
+import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.config.Config.OverwriteMode
 import com.cloudogu.gitops.config.Credentials
-import com.cloudogu.gitops.features.git.GitHandler
+import com.cloudogu.gitops.features.Jenkins
 import com.cloudogu.gitops.git.GitRepo
 import com.cloudogu.gitops.git.GitRepoFactory
 import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
