+++ src/main/groovy/com/cloudogu/gitops/features/argocd/ArgoCD.groovy
@@ -1,8 +1,8 @@
 package com.cloudogu.gitops.features.argocd
 
 import com.cloudogu.gitops.Feature
+import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.features.git.GitHandler
 import com.cloudogu.gitops.git.GitRepoFactory
 import com.cloudogu.gitops.kubernetes.api.HelmClient
 import com.cloudogu.gitops.kubernetes.api.K8sClient
