+++ src/test/groovy/com/cloudogu/gitops/features/ContentLoaderTest.groovy
@@ -1,6 +1,6 @@
 package com.cloudogu.gitops.features
 
-import static ContentLoader.RepoCoordinate
+import static com.cloudogu.gitops.application.content.ContentLoader.RepoCoordinate
 import static com.cloudogu.gitops.config.Config.ContentRepoType
 import static com.cloudogu.gitops.config.Config.ContentSchema.ContentRepositorySchema
 import static com.cloudogu.gitops.config.Config.OverwriteMode
@@ -10,9 +10,10 @@ import static org.mockito.ArgumentMatchers.any
 import static org.mockito.ArgumentMatchers.eq
 import static org.mockito.Mockito.*
 
+import com.cloudogu.gitops.application.content.ContentLoader
+import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.config.Credentials
-import com.cloudogu.gitops.features.git.GitHandler
 import com.cloudogu.gitops.features.git.config.ScmTenantSchema
 import com.cloudogu.gitops.git.GitRepoFactory
 import com.cloudogu.gitops.infrastructure.deployment.DeploymentStrategy
