+++ src/test/groovy/com/cloudogu/gitops/features/git/GitHandlerTest.groovy
@@ -3,6 +3,7 @@ package com.cloudogu.gitops.features.git
 import static org.junit.jupiter.api.Assertions.*
 import static org.mockito.Mockito.mock
 
+import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.features.git.config.util.ScmProviderType
 import com.cloudogu.gitops.git.providers.GitProvider
