+++ src/test/groovy/com/cloudogu/gitops/utils/git/GitHandlerForTests.groovy
@@ -2,8 +2,8 @@ package com.cloudogu.gitops.utils.git
 
 import static org.mockito.Mockito.mock
 
+import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
-import com.cloudogu.gitops.features.git.GitHandler
 import com.cloudogu.gitops.git.providers.GitProvider
 import com.cloudogu.gitops.infrastructure.deployment.HelmStrategy
 import com.cloudogu.gitops.utils.FileSystemUtils
